// src/loans/dto/apply-loan.dto.ts
export class ApplyLoanDto{
    IDCardNumber: string;
  
    IdCardType: string;
  
    RefGhanaNationIdentificationNo: string;
  
    RefLastName: string;
 
    affordability: number;
 
    age: number;
  
    bank_account_holder_name: string;
  
    bank_account_number: string;
  
    bank_branch: string;
  
    bank_name: string;
  
    dateOfBirth: string;
  
    email: string;
  
    employer_name: string;
  
    employment_start_date: string;
  
    facility: string;
  
    firstName: string;
  
    instalmentPerMonth: string;
  
    interestPerMonth: string;
  
    lastName: string;
  
    loanAmount: string;
  
    loanTerm: string;
  
    loantotalCollectable: string;
  
    middleName: string;
  
    modeOfPayment: string;
  
    netSalary: string;
  
    personal_phone: string;
  
    purposeForLoan: string;
  
    refContact1: string;
  
    refContact2: string;
  
    refMiddleName: string;
  
    refSurname: string;
  
    residential_address1: string;
  
    residential_address2: string;
  
    ssnit: string;
  
    staff_id: string;
  
    work_address1: string;
  
    work_address2: string;
  
    work_phone: string;
}